// Angular Frontend/AngularFrontend/src/app/models/employee.model.ts
export interface Employee {
  id: string;
  positions: string[];
  preferredName: string;
  firstName: string;
  lastName: string;
  address: string;
  numOfAvailDaysAWeek: number;
  everydayAvailability: boolean;
  mondayAvailability: boolean;
  tuesdayAvailability: boolean;
  wednesdayAvailability: boolean;
  thursdayAvailability: boolean;
  fridayAvailability: boolean;
  saturdayAvailability: boolean;
  sundayAvailability: boolean;
}